Added basic and crude, but working VR support
Currently supported is TrinusVR (PC->Android streaming). The VR system in the game is on a separate architecural level right above Unity's built-in support and TrinusVR, so that we can accomodate more VR-s without reiventing and changing everything later. VRDevice-s are the low-level implementation of device/plugin/API specific VRs while VRManager is the high level component that establishes the connection between the game and the devices.
VRDeviceTrinusImpl is mostly an altered copy-paste of TrinusProcessor that implements the IVRDevice inteface.

More changes and a lot of refactoring to come soon.